pyproject.toml
[project]
name = "multibagger-stockanalyer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-gpu-cu12>=1.11.0",
    "hydra-core>=1.3.2",
    "pandas>=2.3.0",
    "pip>=25.1.1",
    "py2neo>=2021.2.4",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "sentence-transformers>=4.1.0",
    "spacy>=3.8.7",
    "transformers>=4.52.4",
    "yfinance>=0.2.62",
]
